refactor(events): one save hook, a flatter stage machine, a dedicated watermark module
Three review follow-ups on the watermark refactor, all structural:

- verifyOnSave + syncUnmanagedCheckAt collapse into syncVerificationOnSave.
  They were two branches of one question (what should this save do to the
  verification fields?), mutually exclusive by construction, and each carried
  its own copy of the schedule merge. The skipVerifyHook asymmetry between
  them — the verify branch honours it, the derived-watermark branch must not —
  is now a documented ordering inside one function rather than an invariant
  split across two files.
- StageRule/finishesOnRunOut is gone; MACHINE is Record<VerificationStage,
  StageAction>. The flag encoded a tautology — finishing is a transition
  *into* `finished`, so it cannot apply to an event already there — which
  reads better as a guard at the one call site than as a field repeated across
  eight rows. The invariant it protected is covered end-to-end by the
  retention case in the int spec.
- The watermark rule moves out of periods.ts into its own watermark.ts.
  periods.ts is cadence helpers; the watermark rule is the load-bearing
  contract behind the job's single query, and had turned its host into a grab
  bag.

Also fixes a pre-existing flake in the listing-progress tests (1 of 4 runs):
remindOnce took the first email sent to the manager address that every event
in the spec shares, so a second reminder landing in the same job run could be
asserted against by mistake. It now selects by event title — three
consecutive clean runs after the change.

src/lib/eventVerification stays in lib, deliberately: ten distinct owners
import from it (the Events and Registrations collections, the job, three
admin components, seeds, app routes, two lib folders), so co-locating it
under Events would force cross-owner imports into a collection's internals —
the arrangement .claude/rules/project-structure.md rule 4 prohibits.
